Remove dispatch/attributes.cs
src/evaluator/... has both attributes.cs and dispatch/attributes.cs
(providing fn dispatch_attributes). Combine them to help facilitate
moving evaluate_function_call_ast into core_eval.rs.
